Refactor: mypage를 user로 통합
