+++ src/controllers/user.controller.js
@@ -11,5 +11,18 @@ class UserController {
             return res.status(400).json(errResponse({ isSuccess: false, code: 400, message: error.message }));
         }
     }
+    static async findUsername(req, res){
+        const { email } = req.body;
+        try {
+            const result = await UserService.findUsernameByEmail(email);
+            if (result) {
+            return res.status(201).json(response({ isSuccess: true, code: 201, message: 'Username has been sent to your email' }, result));
+            } else {
+            return res.status(400).json(errResponse({ isSuccess: false, code: 400, message:'Email not found'}));
+            }
+        } catch (error) {
+            res.status(500).send(error.message);
+        }
+    }
 }
 module.exports = UserController;
