+++ src/models/user.dao.js
@@ -42,6 +42,20 @@ class UserDAO {
             });
         });
     }
+    static async getUsernameByEmail(email) {
+        return new Promise((resolve, reject) => {
+            pool.query('SELECT id, nickname, name FROM User WHERE email = ?', [email], (err, results) => {
+              if (err) {
+                return reject(err);
+              }
+              if (results.length > 0) {
+                resolve(results[0]);
+              } else {
+                resolve(null);
+              }
+            });
+        });
+    }
 }
 
 module.exports = UserDAO;
