--- src/routes/mypage.route.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const MypageController = require('../controllers/mypage.controller');
-
-const router = express.Router();
-
-router.post('/find-username', MypageController.findUsername);
-
-module.exports = router;
