+++ src/routes/user.route.js
@@ -4,5 +4,6 @@ const UserController = require('../controllers/user.controller');
 const router = express.Router();
 
 router.post('/signup', UserController.registerUser);
+router.post('/find-username',UserController.findUsername);
 
 module.exports = router;
