--- swagger/mypage.swagger.yaml
@@ -1,83 +1,0 @@
-paths:
-  /mypage/find-username:
-    post:
-      tags:
-        - Mypage
-      summary: 이메일로 아이디 보내주기
-      parameters:
-        - name:
-          in: body
-          required: true
-          schema:
-            properties:
-                email:
-                  type: string
-                  description: 이메일
-                  example: user@naver.com
-      responses:
-        201:
-          description: 아이디 찾기 이메일 보내기 선공
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  status:
-                    type: integer
-                    example: 201
-                  isSuccess:
-                    type: boolean
-                    example: true
-                  code:
-                    type: integer
-                    example: 2000
-                  message:
-                    type: string
-                    example: 'Username has been sent to your email'
-                  data:
-                    type: object
-                    properties:
-                      id:
-                        type: integer
-                        example: 1
-                      email:
-                        type: string
-                        example: user@example.com
-        400:
-          description: 잘못된 요청
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  status:
-                    type: integer
-                    example: 400
-                  isSuccess:
-                    type: boolean
-                    example: false
-                  code:
-                    type: integer
-                    example: COMMON001
-                  message:
-                    type: string
-                    example: 잘못된 요청입니다
-        500:
-          description: 서버 에러
-          content:
-            application/json:
-              schema:
-                type: object
-                properties:
-                  status:
-                    type: integer
-                    example: 500
-                  isSuccess:
-                    type: boolean
-                    example: false
-                  code:
-                    type: integer
-                    example: COMMON000
-                  message:
-                    type: string
-                    example: 서버 에러, 관리자에게 문의 바랍니다.
